feat: add claimEventHour method

+++ src/main/java/com/sep490/g28/hvh/be/dto/eventapplication/response/RegisteredParticipantSimpleResponse.java
@@ -19,7 +19,7 @@ public class  RegisteredParticipantSimpleResponse {
     private String name;
     private String avatarUrl;
     private String address;
-    private Short creditScore;
-    private Short honorScore;
+    private int creditScore;
+    private int honorScore;
     private OffsetDateTime createdAt;
 }
